back-end/requirements.txt
Flask==1.1.2
Flask_Cors==4.0.0
hexbytes==0.3.1
py_solc_x==2.0.2
pycryptodome==3.19.1
pymongo==4.6.3
web3==6.15.1
